% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreal_golem_add_ext.R
\name{app_dashboard}
\alias{app_dashboard}
\title{Exemple app}
\usage{
app_dashboard(browse = FALSE)
}
\arguments{
\item{browse}{Logical. Whether to open code of the example ui}
}
\description{
Exemple app
}
