% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{dr_sm_actionButton}
\alias{dr_sm_actionButton}
\alias{dr_actionButton}
\alias{dr_lg_actionButton}
\title{actionButton}
\usage{
dr_sm_actionButton(inputId, label, icon = NULL, width = NULL, ...,
  type = c("primary", "secondary", "info"))

dr_actionButton(inputId, label, icon = NULL, width = NULL, ...,
  type = c("primary", "secondary", "info"))

dr_lg_actionButton(inputId, label, icon = NULL, width = NULL, ...,
  type = c("primary", "secondary", "info"))
}
\arguments{
\item{inputId}{ID à récupérer côté serveur.}

\item{label}{Texte du bouton.}

\item{icon}{Icon du bouton.}

\item{width}{Taille du bouton.}

\item{...}{Attributs à passer au bouton.}

\item{type}{Type du bouton, pour passer des couleurs peut être l'un des suivants : \code{c("primary", "secondary", "info")}.}
}
\value{
Un objet Shiny contenant un actionButton.
}
\description{
Boutons d'action, comportement similaire à \code{\link[shiny]{actionButton}}.
}
\examples{
if  (interactive()){
   library(shiny)
   ui <- function(request){
     dr_fluidPage(
       h2("plop"),
       tagList(
         dr_sm_actionButton("a", "Un bouton")
       )
     )
   }
   server <- function(input, output, session){
     observe({ print(input$a) })
   }
   shinyApp(ui, server)
}
}
\seealso{
Voir également \code{\link[shiny]{actionButton}}.
}
