% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typo.R
\name{dr_blockquote}
\alias{dr_blockquote}
\title{Citation}
\usage{
dr_blockquote(..., footer = NULL)
}
\arguments{
\item{...}{Contenu de la citation}

\item{footer}{Footer de la citation}
}
\value{
Un éléments Shiny de type blockquote.
}
\description{
Citation
}
