% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{dr_checkboxGroupInput}
\alias{dr_checkboxGroupInput}
\title{Checkbox}
\usage{
dr_checkboxGroupInput(inputId, label = NULL, choices, selected = NULL,
  choiceNames = NULL, type = c("primary", "secondary"))
}
\arguments{
\item{inputId}{id, récupéré côté serveur.}

\item{label}{Titre de la checkbox.}

\item{choices}{Choix possibles}

\item{selected}{Choix sélectionné au départ (part défaut, le premier est pris).}

\item{choiceNames}{Nom des choix, seront récupérés par le serveur. Par défaut, le nom des choix est repris.}

\item{type}{"primary" ou "secondary", couleur de la checkbox.}
}
\value{
Un objet Shiny contenant une checkbox.
}
\description{
Checkbox
}
\examples{
if  (interactive()){
   library(shiny)
   ui <- function(request){
     dr_fluidPage(
       h2("plop"),
       tagList(
         dr_checkboxGroupInput("a", "Une checkbox", choices = letters[1:4]) ,
         dr_checkboxGroupInput("b", "Un Autre", choices = letters[1:4], type = "secondary")
       )
     )
   }
   server <- function(input, output, session){
     observe({ print(input$a) })
     observe({ print(input$b) })
   }
   shinyApp(ui, server)
}
}
\seealso{
Voir également \code{\link[shiny]{checkboxGroupInput}}.
}
