% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daterange.R
\name{dr_dateRangeInput}
\alias{dr_dateRangeInput}
\title{Date Range}
\usage{
dr_dateRangeInput(inputId, label, start = NULL, end = NULL,
  min = NULL, max = NULL, format = "yyyy-mm-dd",
  startview = "month", weekstart = 0, language = "fr",
  separator = " et ", width = NULL, autoclose = TRUE)
}
\arguments{
\item{inputId}{ID de l'input, pour le \code{server}.}

\item{label}{Texte de l'élément.}

\item{start}{Date de début.}

\item{end}{Date de fin.}

\item{min}{Date minimum.}

\item{max}{Date maximum.}

\item{format}{Format de la date.}

\item{startview}{Zoom de départ du calendrier.}

\item{weekstart}{Numéro du jour où commence la semaine.}

\item{language}{Langue.}

\item{separator}{Séparateur du dateRange.}

\item{width}{Taille.}

\item{autoclose}{Fermeture automatique ou non du dateRange.}
}
\value{
Un objet Shiny contenant un dateRange.
}
\description{
Date Range
}
\examples{
if  (interactive()){
   library(shiny)
   ui <- function(request){
     dr_fluidPage(
       h2("plop"),
       tagList(
         dr_dateRangeInput("a", "Choisir une date")
       )
     )
   }
   server <- function(input, output, session){
     observe({ print(input$a) })
   }
   shinyApp(ui, server)
}
}
\seealso{
Voir également \code{\link[shiny]{dateRangeInput}}.
}
