% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{dr_examples}
\alias{dr_examples}
\title{Examples d'apps construites avec le template}
\usage{
dr_examples(which = NULL)
}
\arguments{
\item{which}{Template à afficher.}
}
\description{
Examples d'apps construites avec le template
}
