% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{dr_fluidPage}
\alias{dr_fluidPage}
\title{fluidPage avec les dépendances DREAL}
\usage{
dr_fluidPage(...)
}
\arguments{
\item{...}{Contenu de la page}
}
\value{
Une fluidPage avec les dépendances DREAL.
}
\description{
fluidPage avec les dépendances DREAL
}
