% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{dr_icon}
\alias{dr_icon}
\title{Icones}
\usage{
dr_icon(name = shinydreal::dr_icon_list, size = 1, color = "#5a1e82")
}
\arguments{
\item{name}{Nom de l'icone.}

\item{size}{Taille de l'icone.}

\item{color}{Couleur de l'icone.}
}
\value{
Un objet Shiny avec une îcone.
}
\description{
Icones
}
\examples{
if  (interactive()){
   library(shiny)
   ui <- function(request){
     dr_fluidPage(
       h2("plop"),
       tagList(
         dr_icon("la-jsfiddle"),
         dr_icon("drl-explosion", size = 2),
         dr_icon("drl-cadastre", size = 7, color = 'red')
       )
     )
   }
   server <- function(input, output, session){
   }
   shinyApp(ui, server)
}
}
