% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_icon_list.R
\docType{data}
\name{dr_icon_list}
\alias{dr_icon_list}
\title{Liste des icones}
\format{An object of class \code{character} of length 686.}
\usage{
dr_icon_list
}
\description{
Liste des icones
}
\details{
\itemize{
\item \code{dr_*}: icônes personnalisées DREAL
\item \code{la_*} : icônes de la librairie \href{https://icones8.fr/line-awesome}{Line Awesome}
}
}
\keyword{data}
