% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarpage.R
\name{dr_navbarPage}
\alias{dr_navbarPage}
\title{Page avec navBar}
\usage{
dr_navbarPage(title = "DREAL", menu, body,
  main_class = "container-fluid")
}
\arguments{
\item{title}{Titre de la page.}

\item{menu}{Contenu du menu.}

\item{body}{Corps de la page.}

\item{main_class}{Classe du corps de la page.}
}
\value{
Un objet Shiny.
}
\description{
Page avec navBar
}
