% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{dr_progress}
\alias{dr_progress}
\title{Barre de progression}
\usage{
dr_progress(inputId, type = c("primary", "secondary"), value = 25)
}
\arguments{
\item{inputId}{ID de l'input, pour le \code{server}.}

\item{type}{"primary" ou "secondary"}

\item{value}{Taille de la barre (entre 0 et 100)}
}
\value{
Une barre de progression.
}
\description{
Barre de progression
}
\examples{
if  (interactive()){
   library(shiny)
   ui <- function(request){
     dr_fluidPage(
       h2("plop"),
       tagList(
         dr_progress("a", value = 25), 
         dr_actionButton("go", "go")
       )
     )
   }
   server <- function(input, output, session){
     observeEvent(input$go, {
         update_dr_progress(session, "a", sample(1:100, 1)) 
     })
   }
   shinyApp(ui, server)
}
}
