% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radioButtons.R
\name{dr_radioButton}
\alias{dr_radioButton}
\title{Boutons radio}
\usage{
dr_radioButton(inputId, label = NULL, choices, selected = NULL,
  choiceNames = NULL, type = c("primary", "secondary"))
}
\arguments{
\item{inputId}{ID de l'input, pour le \code{server}.}

\item{label}{Texte de l'élément.}

\item{choices}{Choix possibles.}

\item{selected}{Choix sélectionné au départ (part défaut, le premier est pris).}

\item{choiceNames}{Nom des choix, seront récupérés par le serveur. Par défaut, \code{choices} est repris.}

\item{type}{"primary" ou "secondary", couleur de la checkbox.}
}
\value{
Une liste d'éléments Shiny avec des boutons radio.
}
\description{
Boutons radio permettant une selection unique.
}
\examples{
if  (interactive()){
   library(shiny)
   ui <- function(request){
     dr_fluidPage(
       h2("plop"),
       tagList(
         dr_radioButton("a", "Une checkbox", choices = letters[1:4]) ,
         dr_radioButton("b", "Un Autre", choices = letters[1:4], type = "secondary")
       )
     )
   }
   server <- function(input, output, session){
     observe({ print(input$a) })
     observe({ print(input$b) })
   }
   shinyApp(ui, server)
}
}
