% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{dr_selectInput}
\alias{dr_selectInput}
\title{Select Input}
\usage{
dr_selectInput(inputId, label, choices, selected = NULL)
}
\arguments{
\item{inputId}{ID de l'input, pour le \code{server}.}

\item{label}{Texte de l'élément.}

\item{choices}{Choix possibles.}

\item{selected}{Choix sélectionné au départ (part défaut, le premier est pris).}
}
\description{
Select Input
}
\examples{
if  (interactive()){
   library(shiny)
   ui <- function(request){
     dr_fluidPage(
       h2("plop"),
       tagList(
         dr_selectInput("a", "Une selection", choices= letters[1:5])
       )
     )
   }
   server <- function(input, output, session){
     observe({ print(input$a) })
   }
   shinyApp(ui, server)
}
}
