% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebarPage.R
\name{dr_sidebarItem}
\alias{dr_sidebarItem}
\title{Élément de menu}
\usage{
dr_sidebarItem(id, label)
}
\arguments{
\item{id}{Id du menu, doit correspondre à un id de \code{dr_panel}.}

\item{label}{Texte de l'élément de menu.}
}
\description{
Élément de menu
}
