% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebarPage.R
\name{dr_sidebarPage}
\alias{dr_sidebarPage}
\title{Page avec sidebar}
\usage{
dr_sidebarPage(title = "DREAL", sidebar = tagList(), body)
}
\arguments{
\item{title}{Titre de la page.}

\item{sidebar}{Contenu de la barre de navigation.}

\item{body}{Corps de la page.}
}
\value{
Une page avec une sidebar.
}
\description{
Page avec sidebar
}
