% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeSlider.R
\name{dr_sliderInput}
\alias{dr_sliderInput}
\title{Slider}
\usage{
dr_sliderInput(inputId, min, max, value)
}
\arguments{
\item{inputId}{ID récupéré côté serveur.}

\item{min}{Minimum du slider.}

\item{max}{Maximum du slider.}

\item{value}{Valeur de départ du slider.}
}
\value{
Une tagList avec un sliderInput.
}
\description{
Slider
}
\examples{
if  (interactive()){
   library(shiny)
   ui <- function(request){
     dr_fluidPage(
       h2("plop"),
       tagList(
         dr_sliderInput("a", 1, 10, 5)
       )
     )
   }
   server <- function(input, output, session){
     observe({ print(input$a) })
   }
   shinyApp(ui, server)
}
}
