% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{dr_tableOutput}
\alias{dr_tableOutput}
\title{tableOutput DREAL}
\usage{
dr_tableOutput(outputId, style = c("striped", "bordered"))
}
\arguments{
\item{outputId}{ID de l'output}

\item{style}{"striped" ou "bordered"}
}
\value{
Un output pour server
}
\description{
tableOutput DREAL
}
\examples{
if  (interactive()){
   library(shiny)
   ui <- function(request){
     dr_fluidPage(
       h2("plop"),
       tagList(
         dr_tableOutput("a", "striped"),
         dr_tableOutput("b", "bordered")
       )
     )
   }
   server <- function(input, output, session){
     output$a <- renderTable({
         iris
     })
     output$b <- renderTable({
         mtcars
     })
   }
   shinyApp(ui, server)
}
}
