% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flexdash.R
\name{html_flexdash}
\alias{html_flexdash}
\title{R Markdown Format for Flexible Dashboards with dreal template}
\usage{
html_flexdash(..., css_dir = "libs", offline = FALSE)
}
\arguments{
\item{...}{parameters of \code{\link[flexdashboard]{flex_dashboard}}}

\item{css_dir}{directory relative to Rmd where css is saved}

\item{offline}{Logical.}
}
\description{
R Markdown Format for Flexible Dashboards with dreal template
}
