% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleches.R
\name{dr_arrow_nav}
\alias{dr_arrow_nav}
\alias{dr_square_nav}
\title{Bouton de navigation}
\usage{
dr_arrow_nav(id_left, id_right, value = 0)

dr_square_nav(id_left, id_right, text_left = "Précédent",
  text_right = "Suivant", value = 0)
}
\arguments{
\item{id_left}{ID de la flêche de gauche, à récupérer côté serveur.}

\item{id_right}{ID de la flêche de droite, à récupérer côté serveur.}

\item{value}{valeur de départ des flêches.}

\item{text_left}{Pour \code{dr_square_nav}, texte de l'élément de gauche.}

\item{text_right}{Pour \code{dr_square_nav}, texte de l'élément de droite}
}
\value{
Un objet Shiny avec deux actionButtons.
}
\description{
\code{dr_arrow_nav} se compose de deux boutons simples, \code{dr_square_nav} sont
deux boutons rectangulaires avec du texte.
}
\examples{
if  (interactive()){
   library(shiny)
   ui <- function(request){
     dr_fluidPage(
       h2("plop"),
       tagList(
         dr_arrow_nav("gauche", "droite")
       )
     )
   }
   server <- function(input, output, session){
     observe({ print(input$gauche) })
     observe({ print(input$droite) })
   }
   shinyApp(ui, server)
}
}
