% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onoff.R
\name{dr_onOff}
\alias{dr_onOff}
\alias{dr_sq_onOff}
\title{Boutons on off}
\usage{
dr_onOff(inputId)

dr_sq_onOff(inputId)
}
\arguments{
\item{inputId}{ID à récupérer côté serveur.}
}
\value{
Un objet Shiny.
}
\description{
Boutons on off rectables ou ronds. Côté serveur, renvoie 0 is ON, 1 si OFF.
}
\examples{
if  (interactive()){
   library(shiny)
   ui <- function(request){
     dr_fluidPage(
       h2("plop"),
       tagList(
         dr_onOff("a")
       )
     )
   }
   server <- function(input, output, session){
     observe({ print(input$a) })
   }
   shinyApp(ui, server)
}
}
