% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{update_dr_progress}
\alias{update_dr_progress}
\title{Update de la barre de progression depuis le serveur}
\usage{
update_dr_progress(session, inputId, value)
}
\arguments{
\item{session}{Objet \code{session} de \code{Shiny}}

\item{inputId}{ID de la barre}

\item{value}{Nouvelle valeur de la barre.}
}
\description{
Update de la barre de progression depuis le serveur
}
