% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydreal-package.R
\docType{package}
\name{shinydreal-package}
\alias{shinydreal}
\alias{shinydreal-package}
\title{shinydreal: Template Shiny DREAL}
\description{
Template Shiny pour la DREAL.
}
\author{
\strong{Maintainer}: Colin Fay \email{colin@thinkr.fr}

}
\keyword{internal}
